Add library target to 'rpc' to enable integration testing
Rearrange the module structure slightly by moving 'client.rs' and (the
main fn of) 'server.rs' into 'src/bin', renamed to their actual binary
targets: 'musig-cli' & 'musigd' respectively. This conforms a bit better
to a standard crate layout and allows the explicit '[[bin]]' sections in
Cargo.toml to be removed.

Also turn 'rpc' into a library crate as planned, by adding 'src/lib.rs',
with the (now deduplicated) module hierarchy moved there, out of the
original binary sources which must now import 'rpc' as a separate crate.
This requires some of the modules (namely "rpc::{server,wallet,pb::*}")
to be public, triggering a few Clippy lints in 'wallet' and 'convert'.
Fix the warnings but suppress the less useful 'missing_panics_doc' and
'must_use_candidates' lints globally in Cargo.toml.

Finally, move the logic to spawn a wallet syncing task from the main fn
of 'musigd' to a 'WalletService::spawn_connection' default method, to
minimise the amount of non-library code and avoid its duplication in the
forthcoming integration tests.
