+++ rpc/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod pb {
+    pub mod convert;
+    pub mod musigrpc;
+    pub mod walletrpc;
+}
+
+mod observable;
+mod protocol;
+pub mod server;
+mod storage;
+pub mod wallet;
